requirements.txt
g2p_en==2.1.0
inflect==7.3.1
librosa==0.10.2.post1
matplotlib==3.9.1
networkx==3.3
numba==0.60.0
numpy==2.0.1
pandas==2.2.2
phonemizer==3.2.1
Pillow==10.4.0
pypinyin==0.51.0
PyQt5==5.15.11
PyQt5_sip==12.15.0
pyworld==0.3.4
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.5.1
scipy==1.14.0
tgt==1.5
torch==2.4.0+cu118
tqdm==4.66.2
Unidecode==1.3.8
